v2.4.5.1 -> v2.4.5.2
  - Takanori Kawano: brlock indexing bugfix
  - Ingo Molnar, Jeff Garzik: softirq updates and fixes
  - Al Viro: rampage of superblock cleanups.
  - Jean Tourrilhes: Orinoco driver update v6, IrNET update
  - Trond Myklebust: NFS brown-paper-bag thing
  - Tim Waugh: parport update
  - David Miller: networking and sparc updates
  - Jes Sorensen: m68k update.
  - Ben Fennema: UDF update
  - Geert Uytterhoeven: fbdev logo updates
  - Willem Riede: osst driver updates
  - Paul Mackerras: PPC update
  - Marcelo Tosatti: unlazy swap cache
  - Mikulas Patocka: hpfs update

+++ drivers/net/eepro100.c
@@ -470,7 +470,7 @@ struct speedo_private {
 	struct timer_list timer;			/* Media selection timer. */
 	struct speedo_mc_block *mc_setup_head;/* Multicast setup frame list head. */
 	struct speedo_mc_block *mc_setup_tail;/* Multicast setup frame list tail. */
-	int in_interrupt;					/* Word-aligned dev->interrupt */
+	long in_interrupt;					/* Word-aligned dev->interrupt */
 	unsigned char acpi_pwr;
 	signed char rx_mode;					/* Current PROMISC/ALLMULTI setting. */
 	unsigned int tx_full:1;				/* The Tx queue is full. */
